backend/requirements.txt
# Glunova API — production dependencies
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
python-multipart>=0.0.9
pydantic[email]>=2.9.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
SQLAlchemy>=2.0.36,<3.0.0
psycopg[binary]>=3.2.0,<4.0.0
alembic>=1.14.0,<2.0.0
bcrypt>=4.2.0,<5.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
python-dotenv>=1.0.0
slowapi>=0.1.9,<1.0.0

# Dev / test
pytest>=8.3.0,<9.0.0
httpx>=0.27.0,<1.0.0
